Enable the resolution change API
This is to help the internal team who seem to be requiring the
API for their use. They have been advised about it's limitation
that it does not check for proper usage and they have to be cautious
about using it.

While at it, at least ensure that the global state about initialization
is handled correctly.